Move HCal algorithms to thier final destination
