+++ Geometry/HcalAlgo/plugins/DDHCalTBZposAlgo.cc
@@ -5,13 +5,47 @@
 
 #include <cmath>
 #include <algorithm>
+#include <map>
+#include <string>
+#include <vector>
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/PluginFactory.h"
 #include "DataFormats/Math/interface/GeantUnits.h"
+#include "DetectorDescription/Core/interface/DDTypes.h"
+#include "DetectorDescription/Core/interface/DDAlgorithm.h"
 #include "DetectorDescription/Core/interface/DDLogicalPart.h"
 #include "DetectorDescription/Core/interface/DDCurrentNamespace.h"
 #include "DetectorDescription/Core/interface/DDSplit.h"
-#include "Geometry/HcalAlgo/plugins/DDHCalTBZposAlgo.h"
+#include "DetectorDescription/Core/interface/DDAlgorithmFactory.h"
+
+class DDHCalTBZposAlgo : public DDAlgorithm {
+public:
+  //Constructor and Destructor
+  DDHCalTBZposAlgo();
+  ~DDHCalTBZposAlgo() override;
+
+  void initialize(const DDNumericArguments& nArgs,
+                  const DDVectorArguments& vArgs,
+                  const DDMapArguments& mArgs,
+                  const DDStringArguments& sArgs,
+                  const DDStringVectorArguments& vsArgs) override;
+
+  void execute(DDCompactView& cpv) override;
+
+private:
+  double eta;      //Eta at which beam is focussed
+  double theta;    //Corresponding theta value
+  double shiftY;   //Shift along Y
+  double shiftX;   //Shift along X
+  double zoffset;  //Offset in z
+  double dist;     //Radial distance
+  double tilt;     //Tilt with respect to y-axis
+  int copyNumber;  //Copy Number
+
+  std::string idNameSpace;  //Namespace of this and ALL sub-parts
+  std::string childName;    //Children name
+};
 
 //#define EDM_ML_DEBUG
 using namespace geant_units::operators;
@@ -79,3 +113,5 @@ void DDHCalTBZposAlgo::execute(DDCompactView& cpv) {
                                << mother << " at " << tran << " with " << rot;
 #endif
 }
+
+DEFINE_EDM_PLUGIN(DDAlgorithmFactory, DDHCalTBZposAlgo, "hcal:DDHCalTBZposAlgo");
