+++ Geometry/HcalAlgo/plugins/DDHCalTestBeamAlgo.cc
@@ -5,12 +5,46 @@
 
 #include <cmath>
 #include <algorithm>
+#include <map>
+#include <string>
+#include <vector>
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/PluginFactory.h"
 #include "DataFormats/Math/interface/GeantUnits.h"
+#include "DetectorDescription/Core/interface/DDTypes.h"
+#include "DetectorDescription/Core/interface/DDAlgorithm.h"
 #include "DetectorDescription/Core/interface/DDLogicalPart.h"
 #include "DetectorDescription/Core/interface/DDCurrentNamespace.h"
-#include "Geometry/HcalAlgo/plugins/DDHCalTestBeamAlgo.h"
+#include "DetectorDescription/Core/interface/DDAlgorithmFactory.h"
+
+class DDHCalTestBeamAlgo : public DDAlgorithm {
+public:
+  //Constructor and Destructor
+  DDHCalTestBeamAlgo();
+  ~DDHCalTestBeamAlgo() override;
+
+  void initialize(const DDNumericArguments& nArgs,
+                  const DDVectorArguments& vArgs,
+                  const DDMapArguments& mArgs,
+                  const DDStringArguments& sArgs,
+                  const DDStringVectorArguments& vsArgs) override;
+
+  void execute(DDCompactView& cpv) override;
+
+private:
+  double eta;        //Eta at which beam is focussed
+  double phi;        //Phi    ................
+  double theta;      //Corresponding theta value
+  double distance;   //Distance of the centre of rotation
+  double distanceZ;  //Distance along x-axis of the centre of rotation
+  double dist;       //Overall distance
+  double dz;         //Half length along z of the volume to be placed
+  int copyNumber;    //Copy Number
+
+  std::string idNameSpace;  //Namespace of this and ALL sub-parts
+  std::string childName;    //Children name
+};
 
 //#define EDM_ML_DEBUG
 using namespace geant_units::operators;
@@ -95,3 +129,5 @@ void DDHCalTestBeamAlgo::execute(DDCompactView& cpv) {
                            << ", " << eta << ", " << phi << ")";
 #endif
 }
+
+DEFINE_EDM_PLUGIN(DDAlgorithmFactory, DDHCalTestBeamAlgo, "hcal:DDHCalTestBeamAlgo");
